# api_gateway/api/docs/swagger.yaml
definitions:
...
  forum.CategoryRes:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  forum.Post:
    properties:
      body:
        type: string
      category_id:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  forum.PostCreateReq:
    properties:
      body:
        type: string
      category_id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  forum.PostGetAllRes:
    properties:
      post:
        items:
          $ref: '#/definitions/forum.PostRes'
        type: array
    type: object
  forum.PostRes:
    properties:
      body:
        type: string
      category:
        $ref: '#/definitions/forum.CategoryRes'
      id:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/forum.UserResp'
    type: object
...
  forum.UserResp:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /categories/{category_id}/posts:
    get:
      consumes:
      - application/json
      description: Get all posts by category
      parameters:
      - description: CategoryId
        in: query
        name: category_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts by category
      tags:
      - category
...
  /post:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/forum.PostCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.Post'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - post
  /post/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted
          schema:
            type: string
        "400":
          description: Invalid post ID
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: Get a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostRes'
        "400":
          description: Invalid post ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get post
      tags:
      - post
    put:
      consumes:
      - application/json
      description: Update an existing post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/forum.PostCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.Post'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update post
      tags:
      - post
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts
      tags:
      - post
...
  /posts/search:
    get:
      consumes:
      - application/json
      description: Get all posts
      parameters:
      - description: Title
        in: query
        name: title
        type: string
      - description: Body
        in: query
        name: body
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts
      tags:
      - post
...
  /users/{user_id}/posts:
    get:
      consumes:
      - application/json
      description: Get all posts user
      parameters:
      - description: UserId
        in: query
        name: user_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts by user
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"